% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunctions.R
\name{create_TF_EPInterval}
\alias{create_TF_EPInterval}
\title{Create Encompassing Priors Interval Test Function}
\usage{
create_TF_EPInterval(lower, upper)
}
\arguments{
\item{lower}{The lower end of the interval to be used. Single-sided intervals can be constructed by using \code{-Inf}.}

\item{upper}{The upper end of the interval to be used. Single-sided intervals can be constructed by using \code{Inf}.}
}
\value{
A function that can be passed as the \code{testFunction} argument of, e.g., \code{\link{testHypothesis}}.
}
\description{
Convenience function for using the encompassing priors test function in the case when an interval is to be tested. The null hypothesis is that all of the effect parameters are within the interval.
}
\examples{
\dontrun{
tf = create_TF_EPInterval(-0.5, 0.5)
testHypothesis(prior, post, fact, "f", testFunction = tf)
}
}
