% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{getEffectParameters}
\alias{getEffectParameters}
\title{Calculate Effect Parameter Matrix}
\usage{
getEffectParameters(cellMeans, factors, testedFactors,
  dmFactors = testedFactors, contrastType = NULL, warnOnDrop = FALSE)
}
\arguments{
\item{cellMeans}{A matrix of cell means. Each column is a cell and each row is a sample from the prior or posterior. The columns must correspond to the rows of \code{factors}.}

\item{factors}{See \code{\link{testHypothesis}}.}

\item{testedFactors}{See \code{\link{testHypothesis}}.}

\item{dmFactors}{See \code{\link{testHypothesis}}.}

\item{contrastType}{See \code{\link{testHypothesis}}.}

\item{warnOnDrop}{Emit a warning if columns of the design matrix are dropped due to it not being full rank.}
}
\value{
A matrix of effect parameters where each column is one parameter and each row is a sample from the prior or posterior (depending on what \code{cellmeans} was). The columns are named with factor level names.
}
\description{
From a matrix of cell means, the factors in the experiment, and information about
which factor to test, creates a matrix of ANOVA effect parameters.
}
\details{
The steps of the procedure:
\enumerate{
\item Calculate the design matrix, \code{X}.
1b. If design is not fully crossed, strip excess terms from \code{X}.
\item If \code{mu} are cell means and \code{beta} are effect parameters, then \code{mu = X * beta} and \code{beta = (X' * X)^-1 * X' * mu}. Calculate \code{S = (X' * X)^-1 * X'}.
\item Calculate \code{beta = S * mu}.
\item Select out only the \code{beta} needed for the \code{testedFactors}, \code{beta_s}.
\item Select the corresponding columns (and rows) in \code{X}, \code{X_s}.
\item Complete the required effect parameters with \code{X_s * beta_s}. This is not the same as \code{mu = X * beta} because only subsets of \code{X} and \code{beta} are used.
}
}
\seealso{
See also \code{\link{summarizeEffectParameters}} and \code{\link{groupEffectParameters}}.
}
