% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnwantedFunctions.R
\name{getPartialFilledS}
\alias{getPartialFilledS}
\title{Get Part of S, Filled with Implicit Parameters}
\usage{
getPartialFilledS(factors, testedFactors, dmFactors, contrastType)
}
\arguments{
\item{factors}{See \code{\link{testHypothesis}}.}

\item{testedFactors}{\code{\link{testHypothesis}}.}

\item{dmFactors}{\code{\link{testHypothesis}}.}

\item{contrastType}{May only be one of \code{"contr.sum"}, \code{"contr.treatment"}, or \code{"contr.SAS"}.}
}
\value{
The selected part of the \code{S} matrix, with additional, implicit parameters included.
}
\description{
\code{S = solve(t(X) \%*\% X) \%*\% t(X)}.
Gets part of S, selected with \code{testedFactors}, that has been filled in with information about the implicit effect parameters.
}
