% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{groupEffectParameters}
\alias{groupEffectParameters}
\title{Group Indistinguishable Effect Parameters}
\usage{
groupEffectParameters(postEffects, method = "credInt", CIP = 0.95,
  priorEffects = NULL, equalBF = 3)
}
\arguments{
\item{postEffects}{A matrix of posterior effect parameters, as from \code{\link{getEffectParameters}}.}

\item{method}{The method to use. \code{"credInt"}: Use credible intervals of the difference between effects to determine equality. \code{"BayesFactor"}: Use Bayes factors of a hypothesis test that the difference between effects is 0.}

\item{CIP}{Used if \code{method == "credInt"}. Credible interval proportion.}

\item{priorEffects}{Used if \code{method == "BayesFactor"}. A matrix of prior effects with columns ordered the same as \code{postEffects}.}

\item{equalBF}{Used if \code{method == "BayesFactor"}. If the Bayes factor in favor of the null is greater than equalBF, the effects are considered to be equivalent.}
}
\value{
Invisibly, a list with the following elements:
\itemize{
\item \code{grp}: A matrix containing grouping information. All effects that share a letter are Indistinguishable. This is what is printed.
\item \code{eqm}: A boolean matrix contining information about which effects were equal (indistinguishable). The value \code{TRUE} indicates equality.
\item \code{bf01}: If \code{method == "BayesFactor"}, a matrix containing Bayes factors in favor of the hypothesis that the two effects were equal.
}
}
\description{
Perform pairwise comparisons of effect levels to determine which effects are indistinguishable from one another. By default, this is based on credible intervals. For people who do not want to, or cannot, use Bayes factors estimated with the Savage-Dickey density ratio, this function allows for another method of determining whether effects are present in the data. In addition, this allows for an analysis of which factor levels differ from one another.
}
