% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnwantedFunctions.R
\name{isDesignFullyCrossed}
\alias{isDesignFullyCrossed}
\title{Determine Whether Factorial Design is Fully Crossed}
\usage{
isDesignFullyCrossed(factors, warnOnDuplicate = TRUE)
}
\arguments{
\item{factors}{See \code{\link{testHypothesis}}.}

\item{warnOnDuplicate}{If \code{TRUE} and there are duplicate rows in \code{factors}, a warning that extent is emitted. Note that there are legitmate reasons to have duplicate rows in \code{factors}.}
}
\value{
\code{TRUE} if the design is fully crossed, \code{FALSE} otherwise.
}
\description{
Determine whether the factorial design in \code{factors} is a fully-crossed design.
}
