% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{plotEffectParameterSummary}
\alias{plotEffectParameterSummary}
\title{Plot Effects Parameter Summary}
\usage{
plotEffectParameterSummary(summary, ctName = "mean", varNames = NULL,
  mar = c(3, 6, 1, 1))
}
\arguments{
\item{summary}{A \code{data.frame} such as that returned by \code{\link{summarizeEffectParameters}}.}

\item{ctName}{The name of the column in \code{summary} that contains the measure of central tendency to plot.}

\item{varNames}{A length 2 vector of names of the columns of \code{summary} that contain measures of variability in the form of endpoints of an error bar. If \code{NULL}, it selects the lowest and highest percentiles present in \code{summary}.}

\item{mar}{Passed to the \code{mar} argument of \code{par()}. If \code{NULL}, it is ignored.}
}
\description{
Takes the result of \code{\link{summarizeEffectParameters}} and plots a measure of central tendency (default mean) and variability (default credible interval).
}
