% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summaries.R
\name{summarizeEffectParameters}
\alias{summarizeEffectParameters}
\title{Summary Statistics of Effect Parameters}
\usage{
summarizeEffectParameters(effects, fList = list(mean = mean), ps = c(0.025,
  0.5, 0.975))
}
\arguments{
\item{effects}{A matrix, such as from \code{\link{getEffectParameters}}, where each column is one parameter and each row is an iteration. If the columns are named with the name of the effect, those names will be used.}

\item{fList}{A named list of functions of one vector argument that will be applied to the effects individually.}

\item{ps}{Proportions for which quantiles will be calculated. By default, the median and 95\% credible interval is calculated.}
}
\value{
A \code{data.frame} with summary statistics for each column in \code{effects}.
}
\description{
Calculates summary statistics, like mean, median, and credible interval, of effect parameters.
}
\details{
Note that equality/inequality of effect parameters should not be determined based on their individual credible intervals. You should instead use the credible interval of their difference, which is done by \code{\link{groupEffectParameters}} by default.
}
\seealso{
For plotting the result of this function, \code{\link{plotEffectParameterSummary}}. For grouping indistinguishable effects, \code{\link{groupEffectParameters}}.
}
