% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunctions.R
\name{testFunction_EPA}
\alias{testFunction_EPA}
\title{Test Function: Encompassing Prior}
\usage{
testFunction_EPA(priorEffects, postEffects, I_M0)
}
\arguments{
\item{priorEffects}{A numeric matrix of prior effect parameters. Rows are samples and columns are parameters.}

\item{postEffects}{A numeric matrix of posterior effect parameters. Rows are samples and columns are parameters.}

\item{I_M0}{The indicator function for the constrained model, \code{M_0}. This is a function that takes a vector of effect parameters, checks some constraint on those parameters, and returns 1 (or \code{TRUE}) if the constraint is satisfied or 0 (or \code{FALSE}) if the constraint is not satisfied.}
}
\value{
A list for the following elements:
\itemize{
\item \code{success}: Logical. Whether or not the test was successful. This is \code{FALSE} if the Bayes factors are infinte, NaN, or NA, or \code{TRUE} otherwise.
\item \code{bf10}: The Bayes factor in favor of the encompassing/alternative model, \code{M_1}.
\item \code{bf01}: The Bayes factor in favor of the constrained/null model, \code{M_0}.
\item \code{prior_satisfied}, \code{post_satisfied}: The proportions of the prior and posterior, respectively, that satisfy the constraint of \code{I_M0}. If these are both low and the number of samples satisfying the constraint is small, it indicates that your constraint is possibly too restrictive and that the estimated Bayes factor may be noisy. You should either use a less-restrictive constraint or take more prior and posterior samples.
}
}
\description{
See the "Encompassing Prior Approach" section of the package manual for an explanation of this approach to estimating Bayes factors. There is some encompassing model, \code{M_1}, and a constrained model, \code{M_0}, which is created by placing some constraint on some of the parameters of \code{M_1}. The function \code{I_M0} that is passed to this function determines whether the constraint is satisfied.

Note that you cannot pass this function directly as the \code{testFunction} argument of \code{\link{testHypothesis}} because there is no default value for \code{I_M0}. Thus, you must create a curried function to pass to \code{\link{testHypothesis}} or use \code{\link{create_EPA_intervalTF}}. See the examples for an example of currying.
}
\examples{
\dontrun{
curriedTestFun = function(priorEffects, postEffects) {
  #M_0: Constrain the effects to be less than 2 units from 0
  I_M0 = function(eff) {
    all(abs(eff) < 2)
  }
  testFunction_EPA(priorEffects, postEffects, I_M0)
}

testHypothesis(..., testFunction = curriedTestFun)
}
}
