% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{testHypotheses}
\alias{testHypotheses}
\title{Test Multiple Hypotheses}
\usage{
testHypotheses(prior, post, factors, testedFactors, dmFactors = testedFactors,
  contrastType = NULL, testFunction = testFunction_SDDR,
  usedFactorLevels = NULL, testName = testedFactors)
}
\arguments{
\item{dmFactors}{Character vector (or list) of the same length as \code{testFactors}.}

\item{usedFactorLevels}{List of data frames. If provided, should be the same length as \code{testFactors}.}

\item{testName}{Character vector (or list). An optional name for the tests. If provided, should be the same length as \code{testFactors}.}

\item{testFactors}{Character vector (or list). Interactions should be indicated by putting colons between factor names. For example, the interaction of A and B is given by "A:B". The order of factor names does not matter. If a list, the elements should not be named.}
}
\value{
A \code{data.frame} with one test on each row.
}
\description{
Test Multiple Hypotheses
}
