% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{testHypotheses}
\alias{testHypotheses}
\title{Test Multiple Hypotheses}
\usage{
testHypotheses(prior, post, factors, testedFactors, dmFactors = testedFactors,
  contrastType = NULL, testFunction = testFunction_SDDR,
  usedFactorLevels = NULL, testName = testedFactors)
}
\arguments{
\item{prior}{See \code{\link{testHypothesis}}.}

\item{post}{See \code{\link{testHypothesis}}.}

\item{factors}{See \code{\link{testHypothesis}}.}

\item{testedFactors}{Character vector (or list). Interactions should be indicated by putting colons between factor names. For example, the interaction of A and B is given by "A:B". The order of factor names does not matter. If a list, the elements should not be named.}

\item{dmFactors}{Character vector (or list) of the same length as \code{testedFactors}.}

\item{contrastType}{See \code{\link{testHypothesis}}.}

\item{testFunction}{See \code{\link{testHypothesis}}.}

\item{usedFactorLevels}{List of data frames. If provided, should be the same length as \code{testedFactors}.}

\item{testName}{Character vector (or list). An optional name for the tests. If provided, should be the same length as \code{testedFactors}.}
}
\value{
A \code{data.frame} with one test on each row.
}
\description{
Convenience function for testing multiple hypotheses. E.g., for a two factor design, you could test the main effects of A and B plus the interaction of A and B. If using \code{testHypothesis} you would need to call that function three times, but this function requires only one function call.
}
\examples{
\dontrun{
prior = getPrior()
post = getPost()
factors = data.frame(let = c('A', 'A', 'B', 'B'),
num = c('1', '2', '1', '2'))
# Test both main effects and the interaction
testHypotheses(prior, post, factors, 
  testedFactors = c('let', 'num', 'let:num'))

}
}
